requirements.txt
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
fire==0.3.1
idna==2.10
pycparser==2.20
python-telegram-bot==12.8
redis==3.5.3
requests==2.24.0
six==1.15.0
termcolor==1.1.0
tornado==6.0.4
urllib3==1.25.9
APScheduler==3.6.3
pytz==2020.1
six==1.15.0
tzlocal==2.1
selenium==3.141.0
urllib3==1.25.9
Pillow==7.2.0
feedparser==5.2.1
imgurpython==1.1.7

